Reviewed, but found strange time drift thing if calculate steps another way.

+++ scala/02_modelling/src/main/scala/TauLeap.scala
@@ -20,7 +20,7 @@ trait TauLeap {
       val (prevTime, prevY) = acc.last
 //      val newY = computeNextStep(prevY, nextTime - prevTime)
       val newY = computeNextStep(prevY, stepSize)
-      acc :+ (nextTime, newY)
+      acc :+ (nextTime.toDouble, newY)
     }
   }
 }
